+++ app/src/main/java/com/scipath/becomeaking/adapter/ItemCallback.java
@@ -1,0 +1,7 @@
+package com.scipath.becomeaking.adapter;
+
+import com.scipath.becomeaking.model.item.IItem;
+
+public interface ItemCallback {
+    public void call(IItem item);
+}