+++ app/src/main/java/com/scipath/becomeaking/data/SaveManager.java
@@ -3,6 +3,9 @@ package com.scipath.becomeaking.data;
 import android.content.Context;
 
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.scipath.becomeaking.model.item.IItem;
+import com.scipath.becomeaking.model.item.IItemSerializer;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -17,7 +20,10 @@ public class SaveManager {
 
     public static void saveGame(Context context, GameState gameState) {
         try {
-            String json = new Gson().toJson(gameState);
+            Gson gson = new GsonBuilder()
+                    .registerTypeAdapter(IItem.class, new IItemSerializer())
+                    .create();
+            String json = gson.toJson(gameState);
             try (FileOutputStream fos = context.openFileOutput(FILE_NAME, Context.MODE_PRIVATE)) {
                 fos.write(json.getBytes(StandardCharsets.UTF_8));
             }
@@ -30,7 +36,10 @@ public class SaveManager {
         try {
             try (FileInputStream fis = context.openFileInput(FILE_NAME)) {
                 InputStreamReader isr = new InputStreamReader(fis, StandardCharsets.UTF_8);
-                return new Gson().fromJson(isr, GameState.class);
+                Gson gson = new GsonBuilder()
+                        .registerTypeAdapter(IItem.class, new IItemSerializer())
+                        .create();
+                return gson.fromJson(isr, GameState.class);
             }
         } catch (IOException e) {
             return null;