+++ app/src/main/java/com/scipath/becomeaking/model/StatBonus.java
@@ -18,8 +18,8 @@ public enum StatBonus implements Serializable {
     CostPerDay(R.string.money, R.string.c_d_per_day, R.drawable.icon_coin),
     HealthImpact(R.string.health, R.string.c_d, R.drawable.icon_heart),
     ReputationImpact(R.string.reputation, R.string.c_d, R.drawable.icon_civic_crown),
-    MoneyPerClick(R.string.money, R.string.c_d_per_click, R.drawable.icon_coin);
-
+    MoneyPerClick(R.string.money, R.string.c_d_per_click, R.drawable.icon_coin),
+    Might(R.string.might, R.string.c_d, R.drawable.icon_might);
 
 
     // Individual stat bonus fields