+++ app/src/main/java/com/scipath/becomeaking/model/item/Work.java
@@ -1,24 +1,31 @@
-package com.scipath.becomeaking.model;
+package com.scipath.becomeaking.model.item;
 
 import android.content.Context;
 import android.graphics.drawable.Drawable;
 
 import androidx.appcompat.content.res.AppCompatResources;
 
+import com.scipath.becomeaking.R;
+import com.scipath.becomeaking.model.Personage;
+import com.scipath.becomeaking.model.StatBonus;
+import com.scipath.becomeaking.model.StatBonusesMap;
+
 import java.io.Serializable;
 
 
-public class Item implements Serializable {
+public class Work implements IItem, Serializable {
 
     // Fields
-    private int nameId;
-    private int imageId;
-    private int cost;
-    private boolean bought;
-    private StatBonusesMap statBonuses;
+    protected int nameId;
+    protected int imageId;
+    protected final int interactionNameId = R.id.buy;
+    protected int cost;
+    protected boolean bought;
+    protected StatBonusesMap statBonuses;
+
 
     // Constructors
-    public Item(int nameId, int imageId, int cost) {
+    public Work(int nameId, int imageId, int cost) {
         this.nameId = nameId;
         this.imageId = imageId;
         this.cost = cost;
@@ -26,54 +33,68 @@ public class Item implements Serializable {
         statBonuses = new StatBonusesMap();
     }
 
-    public Item(int nameId, int imageId, int cost, StatBonusesMap statBonuses) {
+    public Work(int nameId, int imageId, int cost, StatBonusesMap statBonuses) {
         this.nameId = nameId;
         this.imageId = imageId;
         this.cost = cost;
         bought = false;
-        this.statBonuses = statBonuses;
-    }
+        this.statBonuses = statBonuses;    }
 
 
     // Accessors
+    @Override
     public int getNameId() {
         return nameId;
     }
 
+    @Override
     public int getImageId() {
         return imageId;
     }
 
+    @Override
+    public int getInteractionNameId() {
+        return interactionNameId;
+    }
+
+    @Override
     public int getCost() {
         return cost;
     }
 
+    @Override
     public boolean isBought() {
         return bought;
     }
 
+    @Override
     public StatBonusesMap getStatBonuses() {
         return statBonuses;
     }
 
 
     // Mutators
+    @Override
     public void setNameId(int nameId) {
         this.nameId = nameId;
     }
 
+    @Override
     public void setImageId(int imageId) {
         this.imageId = imageId;
     }
 
+    @Override
     public void setCost(int cost) {
         this.cost = cost;
     }
 
-    public void setBought(boolean bought) {
-        this.bought = bought;
+    @Override
+    public void setBought(boolean isBought) {
+        this.bought = isBought;
     }
 
+    @Override
     public void setStatBonuses(StatBonusesMap statBonuses) {
         this.statBonuses = statBonuses;
     }
@@ -85,6 +106,8 @@ public class Item implements Serializable {
      *
      * @return The String that contains the item name
      */
+
+    @Override
     public String getName(Context context) {
         return context.getString(nameId);
     }
@@ -94,7 +117,16 @@ public class Item implements Serializable {
      *
      * @return The Drawable that contains the item image
      */
+
+    @Override
     public Drawable getImage(Context context) {
         return AppCompatResources.getDrawable(context, imageId);
+    }
+
+    @Override
+    public boolean interact(Personage personage) {
+        personage.affectHealth(statBonuses.get(StatBonus.HealthImpact));
+        personage.affectReputation(statBonuses.get(StatBonus.ReputationImpact));
+        return true;
     }
 }