+++ app/src/main/java/com/scipath/becomeaking/view/activity/NamingPersonageActivity.java
@@ -15,6 +15,7 @@ import androidx.core.view.WindowInsetsCompat;
 import com.scipath.becomeaking.BecomeAKing;
 import com.scipath.becomeaking.R;
 import com.scipath.becomeaking.data.CategoriesMock;
+import com.scipath.becomeaking.data.GameState;
 import com.scipath.becomeaking.model.Category;
 import com.scipath.becomeaking.model.Personage;
 import com.scipath.becomeaking.model.Sex;
@@ -51,8 +52,7 @@ public class NamingPersonageActivity extends AppCompatActivity {
                 Personage personage = new Personage(personageName, sex, title);
                 ArrayList<Category> categories = CategoriesMock.getCategories(sex == Sex.Male);
                 categories.get(1).recalculateStats(); // To set cloth image
-                BecomeAKing.getInstance().setCurrentPersonage(personage);
-                BecomeAKing.getInstance().setCurrentCategories(categories);
+                BecomeAKing.getInstance().setGameState(new GameState(0, personage, categories));
 
                 Intent intent = new Intent(NamingPersonageActivity.this, GameActivity.class);
                 startActivity(intent);